build(deps)!: update react-datepicker to v6
BREAKING CHANGE: Require date-fns 4
BREAKING CHANGE: bump react-datepicker from v4 to v7
List of breaking changes:
- v5: https://github.com/Hacker0x01/react-datepicker/pull/4393
- v6: https://github.com/Hacker0x01/react-datepicker/pull/4481
- v7: https://github.com/Hacker0x01/react-datepicker/releases/tag/v7.0.0
BREAKING CHANGE: remove includeTime prop in favor of showTimeSelect
BREAKING CHANGE: remove timeLabel prop in favor of timeCaption
Refs: CDS-206
