+++ .storybook/main.ts
@@ -16,7 +16,14 @@ const config: StorybookConfig = {
 		options: {}
 	},
 	typescript: {
-		reactDocgen: 'react-docgen-typescript'
+		reactDocgen: 'react-docgen-typescript',
+		reactDocgenTypescriptOptions: {
+			// Makes string and boolean types that can be undefined appear as inputs and switches
+			shouldRemoveUndefinedFromOptional: true,
+			// Filter out third-party props from node_modules except react-datepicker package
+			propFilter: (prop) =>
+				prop.parent ? !/node_modules\/(?!react-datepicker)/.test(prop.parent.fileName) : true,
+		},
 	},
 	webpackFinal: async (config) => {
 		return {
