+++ jest.config.ts
@@ -101,7 +101,8 @@ const config: Config = {
 	// A map from regular expressions to module names that allow to stub out resources with a single module
 	moduleNameMapper: {
 		'@testing-library/user-event/dist/cjs/keyboard/keyMap':
-			'<rootDir>/node_modules/@testing-library/user-event/dist/cjs/keyboard/keyMap'
+			'<rootDir>/node_modules/@testing-library/user-event/dist/cjs/keyboard/keyMap',
+		'\\.(css|less)$': 'identity-obj-proxy'
 	},
 
 	// An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
