+++ package.json
@@ -67,8 +67,8 @@
 		"@storybook/addon-interactions": "8.3.5",
 		"@storybook/addon-webpack5-compiler-babel": "3.0.3",
 		"@storybook/blocks": "8.3.5",
-        "@storybook/components": "8.3.5",
-        "@storybook/icons": "1.2.12",
+		"@storybook/components": "8.3.5",
+		"@storybook/icons": "1.2.12",
 		"@storybook/react-webpack5": "8.3.5",
 		"@storybook/test": "8.3.5",
 		"@svgr/cli": "^8.1.0",
@@ -79,7 +79,6 @@
 		"@types/jest": "^29.5.13",
 		"@types/lodash": "^4.17.7",
 		"@types/react": "^18.3.7",
-		"@types/react-datepicker": "^4.19.6",
 		"@types/react-dom": "^18.3.0",
 		"@zextras/carbonio-ui-configs": "^1.0.3",
 		"babel-jest": "^29.7.0",
@@ -91,6 +90,7 @@
 		"eslint-plugin-notice": "^1.0.0",
 		"eslint-plugin-storybook": "^0.9.0",
 		"husky": "^9.1.6",
+		"identity-obj-proxy": "^3.0.0",
 		"is-ci": "^3.0.1",
 		"jest": "^29.7.0",
 		"jest-environment-jsdom": "^29.7.0",
@@ -102,10 +102,11 @@
 		"process": "0.11.10",
 		"react-docgen-typescript": "^2.2.2",
 		"rollup": "^4.21.3",
+		"rollup-plugin-postcss": "^4.0.2",
 		"semantic-release": "^24.1.2",
 		"storybook": "8.3.5",
 		"ts-node": "^10 >=10.9.2",
-		"typescript": "^5.4.5"
+		"typescript": "~5.6.3"
 	},
 	"browserslist": [
 		">1%",
@@ -114,6 +115,7 @@
 		"not dead"
 	],
 	"peerDependencies": {
+		"date-fns": "^4.1.0",
 		"lodash": "^4.17.21",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
@@ -124,6 +126,6 @@
 		"core-js": "^3.38.1",
 		"darkreader": "^4.9.92",
 		"polished": "^4.3.1",
-		"react-datepicker": "^4.25.0"
+		"react-datepicker": "file:react-datepicker-7.4.0.tgz"
 	}
 }
