+++ src/assets.d.ts
@@ -1,0 +1,6 @@
+/*
+ * SPDX-FileCopyrightText: 2024 Zextras <https://www.zextras.com>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+declare module '*.css' {}
