+++ src/components/inputs/datetimePicker/DateTimePicker.mdx
@@ -52,7 +52,7 @@ Locale does not change the "time" label, which requires to be translated from th
 The lib react-datepicker works with _date-fns_.
 To avoid importing all locales inside the bundle of the DS, it's left up to the external
 project to install date-fns and configure it to limit the bundle size.
-See [https://date-fns.org/v2.29.3/docs/webpack]() for more details.
+See [https://date-fns.org/v4.1.0/docs/webpack]() for more details.
 
 **Important:** check inside installed dependency date-fns/locale for available locales (e.g.: en does not exists, en-US is the default)
