+++ src/components/inputs/datetimePicker/DateTimePicker.stories.registeredLocale.tsx
@@ -5,7 +5,7 @@
  */
 import React from 'react';
 
-import ptBR from 'date-fns/locale/pt-BR';
+import { ptBR } from 'date-fns/locale/pt-BR';
 
 import { DateTimePicker, registerLocale } from './DateTimePicker';
