+++ src/globals.ts
@@ -14,7 +14,6 @@ interface DSWindow {
 }
 
 declare global {
-	// eslint-disable-next-line @typescript-eslint/no-empty-interface
 	interface Window extends DSWindow {}
 }
