+++ tsconfig.json
@@ -13,7 +13,8 @@
     "skipLibCheck": true,
     "paths": {
       "@testing-library/user-event/dist/cjs/keyboard/keyMap": ["./node_modules/@testing-library/user-event/dist/types/keyboard/keyMap.d.ts"]
-    }
+    },
+    "noUncheckedSideEffectImports": true
   }
 }
